requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
SQLAlchemy>=1.4
psycopg2-binary>=2.9
alembic>=1.10
python-dotenv>=1.0
passlib[bcrypt]>=1.7
PyJWT>=2.8
pydantic>=1.10
python-multipart>=0.0.6
aiofiles>=23.1.0
